Changes after code-review
