+++ pg_view.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from pg_view import main
+
+if __name__ == '__main__':
+    main()
