+++ pg_view/loggers.py
@@ -1,11 +1,11 @@
 import logging
 
-logger = None
+logger = logging.getLogger(__name__)
 
 
 def setup_loggers(options):
     global logger
-    logger = logging.getLogger(__name__)
+
     logger.setLevel((logging.INFO if options.verbose else logging.ERROR))
     log_stderr = logging.StreamHandler()
     logger.addHandler(log_stderr)
