+++ pg_view/meta.py
@@ -1,0 +1,4 @@
+__appname__ = 'pg_view'
+__version__ = '1.3.1'
+__author__ = 'Oleksii Kliukin <oleksii.kliukin@zalando.de>'
+__license__ = 'Apache 2.0'
